// frontend/app/dashboard_student/types.ts
export interface StudentDataFields {
  email: string;
  first_name: string;
  last_name: string;
  parent_id: string;
  age: number;
  grade_level: number;
  sensory_preference: string;
  communication_preference: string;
  attention_span: string;
  reading_writing_skills: string;
  math_skills: string;
  technology_comfort: string;
  interests: string;
}